Consume crossplane-runtime fix for shared informer cache objects
- Remove deprecated MR {Get,Set}ProviderReference methods

+++ apis/redshiftserverless/v1beta1/zz_snapshot_types.go
@@ -16,7 +16,7 @@ import (
 type SnapshotInitParameters struct {
 
 	// How long to retain the created snapshot. Default value is -1.
-	RetentionPeriod *float64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
+	RetentionPeriod *int64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
 }
 
 type SnapshotObservation struct {
@@ -49,7 +49,7 @@ type SnapshotObservation struct {
 	OwnerAccount *string `json:"ownerAccount,omitempty" tf:"owner_account,omitempty"`
 
 	// How long to retain the created snapshot. Default value is -1.
-	RetentionPeriod *float64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
+	RetentionPeriod *int64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
 }
 
 type SnapshotParameters struct {
@@ -75,7 +75,7 @@ type SnapshotParameters struct {
 
 	// How long to retain the created snapshot. Default value is -1.
 	// +kubebuilder:validation:Optional
-	RetentionPeriod *float64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
+	RetentionPeriod *int64 `json:"retentionPeriod,omitempty" tf:"retention_period,omitempty"`
 }
 
 // SnapshotSpec defines the desired state of Snapshot
